rework: threads all the way down (#44)
A pretty sizeable rework to the underlying code but no changes to the API

ALL flushing to Supergood is done in a separate thread. Currently the library uses a synchronous scheme when it detects it is operating in a fork. This was in an effort to optimize data integrity, but it does add additional delay.

Instead, a worker thread is spun up just like it would have been on the main thread. What this means is that sometimes a process that shuts down too quickly after finishing their HTTP calls won't end up flushing to Supergood.

In the future, we intend on utilizing a separately deployable agent that can manage all this off-process, but for now this is my best effort to balance data integrity without compromising speed.

---------

Co-authored-by: Steve Bunting <steve@supergood.ai>